Cursach

+++ 2course2semestr/Cursach/BackEnd/src/main/java/ru/mirea/SidorovSD/Services/GroupService.java
@@ -4,8 +4,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.mirea.SidorovSD.Models.Group;
-import ru.mirea.SidorovSD.Models.Group;
 import ru.mirea.SidorovSD.Repos.GroupRepo;
+import ru.mirea.SidorovSD.Repos.PermissionRepo;
+import ru.mirea.SidorovSD.Repos.UserRepo;
+import ru.mirea.SidorovSD.Repos.VehicleRepo;
 
 import java.util.List;
 
@@ -16,40 +18,48 @@ public class GroupService {
     @Autowired
     private final GroupRepo groupRepo;
 
-    public GroupService(GroupRepo groupRepo) {
+    @Autowired
+    private final PermissionRepo permissionRepo;
+
+    public GroupService(GroupRepo groupRepo, PermissionRepo permissionRepo) {
         this.groupRepo = groupRepo;
+        this.permissionRepo = permissionRepo;
     }
 
     public List<Group> findAll(){
         return groupRepo.findAll();
     }
 
-    public Boolean saveGroup(String groupName){
+    public String saveGroup(String groupName){
         log.info(groupName);
         if (isGroupExist(groupName)){
-            return Boolean.FALSE;
+            return "Group already exist";
         }
         Group group  = new Group();
         group.setGroupName(groupName);
         groupRepo.save(group);
-        return Boolean.TRUE;
+        return "OK";
     }
 
-    public Boolean updateGroup(int idGroup, String groupName){
+    public String updateGroup(int idGroup, String groupName){
         Group group = groupRepo.findByIdGroup(idGroup);
         if (group == null)
-            return false;
+            return "Group doesn't exist";
         group.setGroupName(groupName);
         groupRepo.save(group);
-        return true;
+        return "OK";
     }
 
-    public Boolean deleteGroup(String groupName){
+    public String deleteGroup(String groupName){
         Group group = groupRepo.findByGroupName(groupName);
+
         if (group == null)
-            return false;
+            return "Group doesn't exist";
+        if (!permissionRepo.findByIdGroup(group.getIdGroup()).isEmpty())
+            return "Group in use";
+
         groupRepo.delete(group);
-        return true;
+        return "OK";
     }
 
     public Boolean isGroupExist(String groupName){
