+++ 2course2semestr/Cursach/BackEnd/src/main/java/ru/mirea/SidorovSD/Services/VehicleBrandService.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.mirea.SidorovSD.Models.Vehicle_brand;
 import ru.mirea.SidorovSD.Repos.VehicleBrandRepo;
+import ru.mirea.SidorovSD.Repos.VehicleNameRepo;
 
 import java.util.List;
 
@@ -15,8 +16,12 @@ public class VehicleBrandService {
     @Autowired
     private final VehicleBrandRepo vehicleBrandRepo;
 
-    public VehicleBrandService(VehicleBrandRepo vehicleBrandRepo) {
+    @Autowired
+    private final VehicleNameRepo vehicleNameRepo;
+
+    public VehicleBrandService(VehicleBrandRepo vehicleBrandRepo, VehicleNameRepo vehicleNameRepo) {
         this.vehicleBrandRepo = vehicleBrandRepo;
+        this.vehicleNameRepo = vehicleNameRepo;
     }
 
     public List<Vehicle_brand> getAllBrands(){
@@ -27,31 +32,33 @@ public class VehicleBrandService {
         return vehicleBrandRepo.findByBrandName(brandName);
     }
 
-    public Boolean addBrand(String brandName){
+    public String addBrand(String brandName){
         Vehicle_brand brand = getBrandByName(brandName);
         if (brand != null)
-            return Boolean.FALSE;
+            return "Brand already exist";
         brand = new Vehicle_brand();
         brand.setBrandName(brandName);
         vehicleBrandRepo.save(brand);
-        return Boolean.TRUE;
+        return "OK";
     }
 
-    public Boolean updateBrand(int idBrand, String brandName){
+    public String updateBrand(int idBrand, String brandName){
         Vehicle_brand brand = vehicleBrandRepo.findByIdBrand(idBrand);
         if (brand == null)
-            return Boolean.FALSE;
+            return "Brand already exist";
         brand.setBrandName(brandName);
         vehicleBrandRepo.save(brand);
-        return Boolean.TRUE;
+        return "OK";
     }
 
-    public Boolean deleteBrand(String brandName){
+    public String deleteBrand(String brandName){
         Vehicle_brand brand = getBrandByName(brandName);
         if (brand == null)
-            return false;
+            return "Brand doesn't exist";
+        if (!vehicleNameRepo.findByIdBrand(brand.getIdBrand()).isEmpty())
+            return "Brand in use now in other lists";
         vehicleBrandRepo.delete(brand);
-        return Boolean.TRUE;
+        return "OK";
     }
 
 }
