+++ 2course2semestr/Cursach/BackEnd/src/main/java/ru/mirea/SidorovSD/Services/VehicleNameService.java
@@ -9,6 +9,7 @@ import ru.mirea.SidorovSD.Models.Vehicle_name;
 import ru.mirea.SidorovSD.Repos.VehicleBrandRepo;
 import ru.mirea.SidorovSD.Repos.VehicleModelRepo;
 import ru.mirea.SidorovSD.Repos.VehicleNameRepo;
+import ru.mirea.SidorovSD.Repos.VehicleWorkModelRepo;
 
 import java.util.List;
 
@@ -24,68 +25,101 @@ public class VehicleNameService {
     @Autowired
     private final VehicleModelRepo vehicleModelRepo;
 
-    public VehicleNameService(VehicleNameRepo vehicleNameRepo, VehicleBrandRepo vehicleBrandRepo, VehicleModelRepo vehicleModelRepo) {
+    @Autowired
+    private final VehicleWorkModelRepo vehicleWorkModelRepo;
+
+    public VehicleNameService(VehicleNameRepo vehicleNameRepo, VehicleBrandRepo vehicleBrandRepo, VehicleModelRepo vehicleModelRepo, VehicleWorkModelRepo vehicleWorkModelRepo) {
         this.vehicleNameRepo = vehicleNameRepo;
         this.vehicleBrandRepo = vehicleBrandRepo;
         this.vehicleModelRepo = vehicleModelRepo;
+        this.vehicleWorkModelRepo = vehicleWorkModelRepo;
     }
 
     public List<Vehicle_name> getAllNames(){
         return vehicleNameRepo.findAll();
     }
 
-    public Boolean addName(String brandName, String modelName){
-        Vehicle_name vehicle_name = getName(brandName, modelName);
+    public String addName(String brandName, String modelName){
+
         Vehicle_brand vehicle_brand  = vehicleBrandRepo.findByBrandName(brandName);
         Vehicle_model vehicle_model = vehicleModelRepo.findByModelName(modelName);
-        if (vehicle_brand != null && vehicle_model != null && vehicle_name == null){
-            vehicle_name = new Vehicle_name();
-            vehicle_name.setIdBrand(vehicle_name.getIdBrand());
-            vehicle_name.setIdModel(vehicle_model.getIdModel());
-            vehicleNameRepo.save(vehicle_name);
-            return Boolean.TRUE;
+        if (vehicle_brand != null && vehicle_model != null) {
+            Vehicle_name vehicle_name = getName(brandName, modelName);
+            if (vehicle_name == null){
+                vehicle_name = new Vehicle_name();
+                vehicle_name.setIdBrand(vehicle_brand.getIdBrand());
+                vehicle_name.setIdModel(vehicle_model.getIdModel());
+                vehicleNameRepo.save(vehicle_name);
+                return "OK";
+            }
+            else
+                return "Name doesn't exist";
         }
-        else
-            return Boolean.FALSE;
+        else{
+            return "Brand or Model doesn't exist";
+        }
     }
 
-    public Boolean changeName(int idName, int idBrand, int idModel){
+    public String changeName(int idName, String brandName, String modelName){
         Vehicle_name vehicle_name = getName(idName);
         if (vehicle_name == null)
-            return Boolean.FALSE;
+            return "Name doesn't exist";
 
-        if (idBrand != -1) {
-            if (checkIfBrandExist(idBrand))
-                vehicle_name.setIdBrand(idBrand);
+        if (!brandName.equals("-")) {
+            Vehicle_brand brand = vehicleBrandRepo.findByBrandName(brandName);
+            if (brand != null)
+                vehicle_name.setIdBrand(brand.getIdBrand());
             else
-                return false;
+                return "Brand name doesn't exist";
         }
 
-        if(idModel != -1){
-            if (checkIfModelExist(idModel))
-                vehicle_name.setIdModel(idModel);
+        if(!modelName.equals("-")){
+            Vehicle_model model = vehicleModelRepo.findByModelName(modelName);
+            if (model != null)
+                vehicle_name.setIdModel(model.getIdModel());
             else
-                return false;
+                return "Model name doesn't exist";
         }
-
-        return Boolean.TRUE;
+        vehicleNameRepo.save(vehicle_name);
+        return "OK";
 
     }
 
-    public Boolean deleteName(int idName){
+    public String deleteName(int idName){
         Vehicle_name vehicle_name = getName(idName);
         if (vehicle_name == null)
-            return false;
+            return "Name doesn't exist";
+        if (!vehicleWorkModelRepo.findByIdVehicleName(idName).isEmpty())
+            return "Name in user";
         vehicleNameRepo.delete(vehicle_name);
-        return true;
+        return "OK";
+    }
+
+    public String deleteName(String brandName, String modelName){
+
+        Vehicle_name vehicle_name = getName(brandName,modelName);
+        if (vehicle_name == null)
+            return "Name doesn't exist";
+        if (!vehicleWorkModelRepo.findByIdVehicleName(vehicle_name.getIdVehicleName()).isEmpty())
+            return "Name in user";
+        vehicleNameRepo.deleteById(vehicle_name.getIdVehicleName());
+        return "OK";
     }
 
     public Boolean checkIfBrandExist(int idBrand){
         return vehicleBrandRepo.existsById(idBrand);
     }
 
+    public Boolean checkIfBrandExist(String brandName){
+        return vehicleBrandRepo.findByBrandName(brandName) != null;
+    }
+
     public Boolean checkIfModelExist(int idModel){
         return vehicleModelRepo.existsById(idModel);
+    }
+
+    public Boolean checkIfModelExist(String modelName){
+        return vehicleModelRepo.findByModelName(modelName) != null;
     }
 
     public Vehicle_name getName(int idBrand, int idModel){
@@ -93,8 +127,10 @@ public class VehicleNameService {
     }
 
     public Vehicle_name getName(String brandName, String modelName){
-
-        return vehicleNameRepo.findByIdBrandAndIdModel(vehicleBrandRepo.findByBrandName(brandName).getIdBrand(), vehicleModelRepo.findByModelName(modelName).getIdModel());
+        Vehicle_brand brand = vehicleBrandRepo.findByBrandName(brandName);
+        Vehicle_model model = vehicleModelRepo.findByModelName(modelName);
+        if(brand == null || model == null) return null;
+        return vehicleNameRepo.findByIdBrandAndIdModel(brand.getIdBrand(), model.getIdModel());
     }
 
     public Vehicle_name getName(int idName){
